fastchat_demo/requirements.txt
python-dotenv>=0.21.0
zhipuai>=2.0.1
fschat>=0.2.35
langchain-openai>=0.0.5
